[#188] refactor: 팀 생성 폴더구조 변경
