+++ src/components/TeamMaking/Buttons/PlusButton/Styles.ts
